Home page (#2)
* chore: :construction: setting-landPage

* feat: :rotating_light: add-Linter-editorConfig-Prettier

* feat: :sparkles: tailwindWorking

* feat: :sparkles: addHeader
