+++ src/pages/Home.jsx
@@ -1,0 +1,12 @@
+import Header from '../components/Header/Header';
+
+function Home() {
+  return (
+    <>
+      <Header />
+      {/* <h2>Esto es el Home</h2> */}
+    </>
+  );
+}
+
+export default Home;
